refactor: Tasklist part of C8 single application
