+++ operate/common/src/main/java/io/camunda/operate/JacksonConfig.java
@@ -22,9 +22,11 @@ import io.camunda.operate.connect.OperateDateTimeFormatter;
 import java.time.Instant;
 import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 
 @Configuration
 public class JacksonConfig {
@@ -65,5 +67,11 @@ public class JacksonConfig {
   @Bean
   public DateTimeFormatter dateTimeFormatter(final OperateDateTimeFormatter dateTimeFormatter) {
     return dateTimeFormatter.getGeneralDateTimeFormatter();
+  }
+
+  @Bean
+  public MappingJackson2HttpMessageConverter mappingJackson2HttpMessageConverter(
+      @Qualifier("operateObjectMapper") final ObjectMapper objectMapper) {
+    return new MappingJackson2HttpMessageConverter(objectMapper);
   }
 }
