+++ operate/data-generator/src/main/java/io/camunda/operate/data/AbstractDataGenerator.java
@@ -25,13 +25,18 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.DependsOn;
 
 @DependsOn("schemaStartup")
 public abstract class AbstractDataGenerator implements DataGenerator {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(AbstractDataGenerator.class);
-  @Autowired protected ZeebeClient client;
+
+  @Autowired
+  @Qualifier("zeebeClient")
+  protected ZeebeClient client;
+
   @Autowired protected OperateProperties operateProperties;
   protected boolean manuallyCalled = false;
   protected ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(3);
