+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/util/j5templates/OperateSearchAbstractIT.java
@@ -27,6 +27,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.security.test.context.support.WithMockUser;
@@ -60,7 +61,10 @@ public class OperateSearchAbstractIT {
   @Autowired protected TestSearchRepository testSearchRepository;
   @Autowired protected SearchContainerManager searchContainerManager;
   @Autowired protected TestResourceManager testResourceManager;
-  @Autowired protected ObjectMapper objectMapper;
+
+  @Autowired
+  @Qualifier("operateObjectMapper")
+  protected ObjectMapper objectMapper;
 
   @BeforeAll
   public void beforeAllSetup() throws Exception {
