+++ operate/schema/src/main/java/io/camunda/operate/store/elasticsearch/ElasticsearchUserStore.java
@@ -29,6 +29,7 @@ import org.elasticsearch.xcontent.XContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.context.annotation.Profile;
@@ -52,7 +53,9 @@ public class ElasticsearchUserStore implements UserStore {
 
   @Autowired protected RestHighLevelClient esClient;
 
-  @Autowired protected ObjectMapper objectMapper;
+  @Autowired
+  @Qualifier("operateObjectMapper")
+  protected ObjectMapper objectMapper;
 
   @Autowired private UserIndex userIndex;
