+++ tasklist/importer-860/src/main/java/io/camunda/tasklist/zeebeimport/v860/processors/common/UserTaskRecordToTaskEntityMapper.java
@@ -24,6 +24,7 @@ import java.time.OffsetDateTime;
 import java.util.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -45,7 +46,8 @@ public class UserTaskRecordToTaskEntityMapper {
   private final ObjectMapper objectMapper;
 
   public UserTaskRecordToTaskEntityMapper(
-      final ObjectMapper objectMapper, final FormStore formStore) {
+      @Qualifier("tasklistObjectMapper") final ObjectMapper objectMapper,
+      final FormStore formStore) {
     this.objectMapper = objectMapper;
     this.formStore = formStore;
   }
