+++ tasklist/importer-common/src/main/java/io/camunda/tasklist/zeebeimport/util/XMLUtil.java
@@ -18,21 +18,17 @@ import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParserFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
 @Component
-@Configuration
 public class XMLUtil {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(XMLUtil.class);
 
-  @Bean
-  public SAXParserFactory getSAXParserFactory() {
+  private SAXParserFactory getSAXParserFactory() {
     final SAXParserFactory saxParserFactory = SAXParserFactory.newInstance();
     saxParserFactory.setNamespaceAware(true);
     try {
