+++ tasklist/importer/src/main/java/io/camunda/tasklist/ImportModuleConfiguration.java
@@ -13,11 +13,17 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.FullyQualifiedAnnotationBeanNameGenerator;
 
 @Configuration
 @ComponentScan(
     basePackages = "io.camunda.tasklist.zeebeimport",
+    excludeFilters = {
+      @ComponentScan.Filter(
+          type = FilterType.REGEX,
+          pattern = "io\\.camunda\\.tasklist\\.zeebeimport\\.security\\..*")
+    },
     nameGenerator = FullyQualifiedAnnotationBeanNameGenerator.class)
 @ConditionalOnProperty(
     name = "camunda.tasklist.importerEnabled",
