+++ tasklist/qa/integration-tests/src/test/java/io/camunda/tasklist/util/apps/modules/ModulesTestApplication.java
@@ -8,11 +8,15 @@
 package io.camunda.tasklist.util.apps.modules;
 
 import io.camunda.tasklist.TasklistModuleConfiguration;
+import io.camunda.tasklist.archiver.security.ArchiverSecurityModuleConfiguration;
 import io.camunda.tasklist.data.DataGenerator;
 import io.camunda.tasklist.data.es.DevDataGeneratorElasticSearch;
 import io.camunda.tasklist.data.os.DevDataGeneratorOpenSearch;
 import io.camunda.tasklist.util.TasklistZeebeIntegrationTest;
 import io.camunda.tasklist.util.TestApplication;
+import io.camunda.tasklist.webapp.management.WebappManagementModuleConfiguration;
+import io.camunda.tasklist.webapp.security.WebappSecurityModuleConfiguration;
+import io.camunda.tasklist.zeebeimport.security.ImporterSecurityModuleConfiguration;
 import io.camunda.webapps.WebappsModuleConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -49,7 +53,13 @@ import org.springframework.context.annotation.Import;
           value = TasklistModuleConfiguration.class)
     },
     nameGenerator = FullyQualifiedAnnotationBeanNameGenerator.class)
-@Import(WebappsModuleConfiguration.class)
+@Import({
+  WebappsModuleConfiguration.class,
+  WebappSecurityModuleConfiguration.class,
+  ArchiverSecurityModuleConfiguration.class,
+  ImporterSecurityModuleConfiguration.class,
+  WebappManagementModuleConfiguration.class,
+})
 public class ModulesTestApplication {
 
   public static void main(final String[] args) throws Exception {
