+++ tasklist/qa/integration-tests/src/test/resources/config/application-test.yml
@@ -32,6 +32,12 @@ spring:
       static-locations: classpath:/test-static/tasklist/
   thymeleaf:
     prefix: classpath:/test-static/
+server.servlet.session.cookie.name: TASKLIST-SESSION
+server.error.include-message: always
+management.health.defaults.enabled: false
+management.endpoint.health.probes.enabled: true
+management.endpoints.web.exposure.include: health,prometheus,loggers,usage-metrics,backups
+management.endpoint.health.group.readiness.include: readinessState,searchEngineCheck
 ---
 spring.config.activate.on-profile: identity-auth, sso-auth
 camunda.webapps.login-delegated: true
