+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/security/TasklistProfileService.java
@@ -7,53 +7,21 @@
  */
 package io.camunda.tasklist.webapp.security;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.Set;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
-import org.springframework.stereotype.Component;
 
-@Component
-public class TasklistProfileService {
+public interface TasklistProfileService {
 
-  public static final String SSO_AUTH_PROFILE = "sso-auth";
-  public static final String IDENTITY_AUTH_PROFILE = "identity-auth";
-  public static final String AUTH_PROFILE = "auth";
-  public static final String DEFAULT_AUTH = AUTH_PROFILE;
-  public static final String LDAP_AUTH_PROFILE = "ldap-auth";
-  public static final Set<String> AUTH_PROFILES =
+  String SSO_AUTH_PROFILE = "sso-auth";
+  String IDENTITY_AUTH_PROFILE = "identity-auth";
+  String AUTH_PROFILE = "auth";
+  String DEFAULT_AUTH = AUTH_PROFILE;
+  String LDAP_AUTH_PROFILE = "ldap-auth";
+  Set<String> AUTH_PROFILES =
       Set.of(AUTH_PROFILE, LDAP_AUTH_PROFILE, SSO_AUTH_PROFILE, IDENTITY_AUTH_PROFILE);
 
-  private static final Set<String> CANT_LOGOUT_AUTH_PROFILES = Set.of(SSO_AUTH_PROFILE);
+  String getMessageByProfileFor(Exception exception);
 
-  @Autowired private Environment environment;
+  boolean currentProfileCanLogout();
 
-  public String getMessageByProfileFor(final Exception exception) {
-    if (isDevelopmentProfileActive()) {
-      return exception.getMessage();
-    }
-    return "";
-  }
-
-  public boolean currentProfileCanLogout() {
-    return Arrays.stream(environment.getActiveProfiles())
-        .noneMatch(CANT_LOGOUT_AUTH_PROFILES::contains);
-  }
-
-  public boolean isDevelopmentProfileActive() {
-    return List.of(environment.getActiveProfiles()).contains("dev");
-  }
-
-  public boolean isSSOProfile() {
-    return Arrays.asList(environment.getActiveProfiles()).contains(SSO_AUTH_PROFILE);
-  }
-
-  public boolean isIdentityProfile() {
-    return Arrays.asList(environment.getActiveProfiles()).contains(IDENTITY_AUTH_PROFILE);
-  }
-
-  public boolean isLoginDelegated() {
-    return isIdentityProfile() || isSSOProfile();
-  }
+  boolean isLoginDelegated();
 }
